+++ Source/ACE/Entity/WorldPackages/WolrdObjectPackage.cs
@@ -15,8 +15,8 @@ namespace ACE.Entity.WorldPackages
 {
     public class WolrdObjectPackage
     {
-        public ObjectGuid Guid { get; }
         public ObjectType Type { get; }
+        public ObjectGuid Guid { get; }
         public virtual float ListeningRadius { get; protected set; } = 0f;
 
         //not sure what to do about this yet...
@@ -51,15 +51,11 @@ namespace ACE.Entity.WorldPackages
         //todo: // make this work..
         public void Render(System.IO.BinaryWriter writer)
         {
-            //var objectCreate = new ServerPacket(0x18, PacketHeaderFlags.EncryptedChecksum);
-            //var objectCreateFragment = new ServerPacketFragment(0x0A, GameMessageOpcode.ObjectCreate);
-            //objectCreateFragment.Payload.WriteGuid(Guid);
-
+            writer.Write((uint)100);
             WorldObjectSegmentModelData.Render(writer);
             WorldObjectSegmentPhysicsData.Render(writer);
             WolrdObjectSegmentGameData.Render(writer);
         }
 
     }
-
-    }
+}
